requirements.txt
streamlit
tensorflow
opencv-python
numpy
pandas
pillow
matplotlib
